+++ contracts/interfaces/ICoverToken.sol
@@ -7,18 +7,18 @@ pragma solidity >=0.6.0 <0.8.0;
  */
 interface ICoverToken {
 
-  function expirationTimestamp() external view returns (uint48);
-  function collateral() external view returns (address);
-  function name() external view returns (string memory);
-  function claimNonce() external view returns (uint256);
+    function expirationTimestamp() external view returns (uint48);
+    function collateral() external view returns (address);
+    function name() external view returns (string memory);
+    function claimNonce() external view returns (uint256);
 
-  function redeemClaim() external;
-  function redeemNoclaim() external;
-  function redeemCollateral(uint256 _amount) external;
+    function redeemClaim() external;
+    function redeemNoclaim() external;
+    function redeemCollateral(uint256 _amount) external;
 
-  /// @notice access restriction - owner (Protocol)
-  function mint(uint256 _amount, address _receiver) external;
+    /// @notice access restriction - owner (Protocol)
+    function mint(uint256 _amount, address _receiver) external;
 
-  /// @notice access restriction - dev
-  function setCovTokenSymbol(string calldata _name) external;
+    /// @notice access restriction - dev
+    function setCovTokenSymbol(string calldata _name) external;
 }
