+++ contracts/interfaces/ILendingPoolAddressesProvider.sol
@@ -4,5 +4,5 @@ pragma solidity >=0.6.0 <0.8.0;
 
 // Aave LendingPool interface
 interface ILendingPoolAddressesProvider {
-	function getLendingPool() external view returns (address);
+    function getLendingPool() external view returns (address);
 }
